Update the application's bundle
Update most of the gems (and lock some versions):

  - Rails (3.2.14)
  - Redcarpet (3.0.0)
  - Draper (1.2.1)
  - Slim (2.0.0)
  - ActiveAdmin (0.5.0, not latest version)
  - Nokogiri (1.6.0)
  - ActiveAttr (0.8.2)

Also disable ActiveAdmin comments since they are useless and break my
balls. Generate a migration to remove the tables and set the comments
option to false in the gem's initializer.

Finally update Article and Comment decorators to call model instead of
source since we've updated to Draper 1.2.1.