Make articles commentable and un-commentable
Allow to disable comments on certain articles. This commit adds a
commentable column to the articles table which is a boolean with true as
its default value.

When its value its true, the behavior stay the same as previously (any
one can post a comment on the article). Otherwise, when the field is set
to false, a notice is displayed instead of the form to inform the user
that comments are disabled.

Also update the resource's decorator not to display a link with anchors
to the comments in case comments are disabled.